requirements.txt
asgiref==3.7.2
boto3==1.36.25
botocore==1.36.25
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.0.0
Django==4.2.3
django-heroku==0.3.1
django-import-export==3.2.0
django-s3-sqlite==0.0.5
django-storages==1.14.5
et-xmlfile==1.1.0
gunicorn==21.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupPy==1.14
MarkupSafe==2.1.3
numpy==1.25.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pillow==11.1.0
psycopg2==2.9.6
pyodbc==5.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.11.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.3.0
whitenoise==6.5.0
xlrd==2.0.1
XlsxWriter==3.1.3
xlwt==1.3.0
